refactor: move lib/gui/StyleUtils.h => apps/deskflow-gui/StyleUtils.h
